requirements.txt
Flask==1.1.2
google-api-core==1.26.0
google-api-python-client==1.12.8
google-auth==1.26.1
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
httplib2==0.18.1
line-bot-sdk==1.18.0
mutagen==1.45.0
oauth2client==4.1.3
PyDrive==1.3.1
requests==2.25.1
youtube-dl==2021.2.10
python-dotenv==0.15.0